update jeval.py. update short-gen.py. bugfix FFModel.
* jeval.py: allow passing in modeldir from commandline.
* jeval.py: bugfix get_missing_models to use realpath.
* short-gen.py: add pass to generate compliment language classes.
* short-gen.py: don't overwrite files by default. add force option.
* FFModel.from_path: default fcnts to 0 if not present in json.

+++ bin/short-gen.py
@@ -9,6 +9,7 @@ import pandas as pd
 
 from src.core.context import Context
 from src.core.app import harness
+from src.data.flexfringe import FFData
 from src.data.mlrt import MLRT_DIR, MLRegTestFile
 
 
@@ -134,10 +135,29 @@ def get_short_data_from_ltag(ltag: str):
     )
 
 
+def generate_compliment(path: str) -> str:
+    ff_data = FFData.from_path(path)
+    # Flip samples
+    header, samples = ff_data.header, []
+    for label, length, seq in ff_data.samples:
+        assert label in (0, 1)
+        assert length == len(seq)
+        label = 0 if label == 1 else 1
+        samples.append((label, length, seq))
+    assert header == ff_data.header
+    assert len(samples) == len(ff_data.samples)
+    # Convert to string.
+    lines = []
+    lines.append(" ".join(map(str, header)))
+    for smpl in samples:
+        lines.append(" ".join(map(str, [smpl[0], smpl[1]] + smpl[2])))
+    return "\n".join(lines)
+
+
 def main(ctx: Context) -> None:
     ctx.parser.add_argument("-o", "--outdir", required=True)
+    ctx.parser.add_argument("-f", "--force", action="store_true", help="overwrite existing")
     args = ctx.parser.parse_args()
-
 
     for data_size in ("Small", "Mid", "Large"):
         outdir = os.path.join(args.outdir, data_size)
@@ -147,14 +167,33 @@ def main(ctx: Context) -> None:
             split = os.path.basename(path).split("_")[1].split(".")[0]
             assert split == "Train"
             if "co" in ltag:
-                continue  # XXX: No fst files for these?
+                continue  # No fst files for these. Generate on second pass.
             outpath = os.path.join(outdir, f"{ltag}_TrainPS.txt")
+            if os.path.exists(outpath) and not args.force:
+                continue  # Skip existing files.
             og_data = list(pd.read_csv(
                 path, sep="\t", names=["sample", "label"]
             ).itertuples(index=False, name=None))
             data = get_short_data_from_ltag(ltag) + og_data
             pd.DataFrame(data).to_csv(outpath, index=False, header=False, sep="\t")
             ff_string = MLRegTestFile.from_path(outpath).to_string()
+            with open(outpath, "w") as fd:
+                fd.write(ff_string)
+            ctx.log.info("wrote: %s", outpath)
+        # Generate compliment classes on second pass.
+        for path in glob(os.path.join(MLRT_DIR, data_size, "*Train.txt")):
+            ltag = os.path.basename(path).split("_")[0]
+            split = os.path.basename(path).split("_")[1].split(".")[0]
+            assert split == "Train"
+            if "co" not in ltag:
+                continue  # Only generate for compliment classes.
+            outpath = os.path.join(outdir, f"{ltag}_TrainPS.txt")
+            if os.path.exists(outpath) and not args.force:
+                continue  # Skip existing files.
+            ff_string = generate_compliment(
+                inpath := outpath.replace(ltag, ltag.replace("co", ""))
+            )
+            ctx.log.info("read: %s", inpath)
             with open(outpath, "w") as fd:
                 fd.write(ff_string)
             ctx.log.info("wrote: %s", outpath)
